fix(mcp): break circular import between googApiKeyAuth.ts and auth.ts (#9297)
Extract readHeaderValue() and AuthRequestHeaders type to a shared
headerReader.ts module so both files import from there instead of
from each other. This breaks the circular dependency that caused
esbuild to emit an invalid __esm wrapper for googApiKeyAuth.ts
(await inside a non-async function), producing a SyntaxError when
running omniroute --mcp.

Closes #9297
