+++ changelog.d/fixes/9297-fix.plan.md
@@ -1,0 +1,1 @@
+- fix(mcp): break circular import between googApiKeyAuth.ts and auth.ts to fix esbuild SyntaxError in MCP server bundle (#9297)
