+++ src/sse/services/googApiKeyAuth.ts
@@ -1,6 +1,4 @@
-import { readHeaderValue } from "./auth.ts";
-
-type AuthRequestHeaders = Headers | Record<string, string | string[] | undefined>;
+import { readHeaderValue, type AuthRequestHeaders } from "./headerReader.ts";
 
 /**
  * Issue #7034: `gemini-cli` (and any `@google/genai`-based client) sends its
